backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.10.4
pydantic-settings==2.7.0
email-validator==2.2.0
python-dotenv==1.0.1
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.14.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
httpx==0.27.2
langchain==0.2.16
langchain-groq==0.1.10
pytest==8.3.4
